+++ docs/docs.go
@@ -793,7 +793,7 @@ const docTemplate = `{
 
 // SwaggerInfo holds exported Swagger Info so clients can modify it
 var SwaggerInfo = &swag.Spec{
-	Version:          "0.0.11-alpha", // 0.1.0
+	Version:          "0.1.0", // "0.0.11-alpha",
 	Host:             "localhost:8080",
 	BasePath:         "/api/v1",
 	Schemes:          []string{},